requirements.txt
langchain==0.3.26
pypdf==5.6.1 
python-dotenv==1.1.0
langchain-community==0.3.26
langchain-pinecone==0.2.8
faiss-cpu
streamlit
langchain-groq==0.2.0 
groq==0.13.1
sentence-transformers==4.1.0